+++ pcl/pcsfont.c
@@ -400,8 +400,8 @@ bitmap:     pfont = gs_alloc_struct(mem, gs_font_base, &st_gs_font_base,
             if ( code < 0 )
               return code;
             pl_fill_in_tt_font(pfont, NULL, gs_next_ids(mem, 1));
-            { uint pitch_cp =
-                pl_get_uint16(pfh->Pitch) * 1000.0 / pfont->data.unitsPerEm;
+            { uint pitch_cp = (uint)(pl_get_uint16(pfh->Pitch) * 1000.0 /
+                                     pfont->data.unitsPerEm);
               pl_fp_set_pitch_cp(&plfont->params, pitch_cp);
             }
           }
@@ -419,8 +419,8 @@ bitmap:     pfont = gs_alloc_struct(mem, gs_font_base, &st_gs_font_base,
             if ( code < 0 )
               return code;
             pl_fill_in_intelli_font(pfont, gs_next_ids(mem, 1));
-            { uint pitch_cp =
-                pl_get_uint16(pfh->Pitch) * 1000.0 / 8782.0;
+            { uint pitch_cp = (uint)(pl_get_uint16(pfh->Pitch) * 1000.0 /
+                                     8782.0);
               pl_fp_set_pitch_cp(&plfont->params, pitch_cp);
             }
             break;
